+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Release TBD
   * Move payment gateway management functions to a separate class.
   * Add syndication classes to support catalog feeds.
   * Deprecate updating cart currency, just purge carts if currency is changed.
+  * Allow sites to define where each address field is required.
 
 ## v1.4.1
 Release 2022-04-07